Fix `test_zipfile` failure when `SOURCE_DATE_EPOCH` is set
Additionally, move the `SOURCE_DATE_EPOCH` test helpers from `test_py_compile` to `test.support.os_helper`.
